Refactor separator file generate component
